+++ gladis/webhooks/github_parser.py
@@ -4,11 +4,14 @@ from core.helpers import str_to_bool
 
 
 class GithubParser:
-    
-    def send_slack_message_to_user(self, message, github_username, send_slack_message=True):
+    def send_slack_message_to_user(
+        self, message, github_username, send_slack_message=True
+    ):
         if send_slack_message:
-            SlackClient().send_slack_direct_message(message, SlackClient.get_slack_username(github_username))
-            
+            SlackClient().send_slack_direct_message(
+                message, SlackClient.get_slack_username(github_username)
+            )
+
     def send_slack_message_to_channel(self, message, send_slack_message=True):
         if send_slack_message:
             SlackClient().send_slack_message_to_channel(message)
@@ -35,7 +38,7 @@ class GithubParser:
 
         merged = payload.get("pull_request", {}).get("merged")
         merged = str_to_bool(merged)
-        
+
         action = payload.get("action")
         if action and action == "opened":
             slack_message = SlackClient.add_to_slack_string(
@@ -43,7 +46,10 @@ class GithubParser:
             )
         elif action and action == "closed":
             if merged:
-                self.send_slack_message_to_user("Congrats! Your PR was merged! :tada:", payload.get("pull_request", {}).get("user", {}).get("login"))
+                self.send_slack_message_to_user(
+                    "Congrats! Your PR was merged! :tada:",
+                    payload.get("pull_request", {}).get("user", {}).get("login"),
+                )
                 slack_message = SlackClient.add_to_slack_string(
                     slack_message, "PR merged! :merged_parrot:"
                 )
@@ -52,7 +58,6 @@ class GithubParser:
                     slack_message, "PR closed! :sad:"
                 )
         SlackClient.add_to_slack_string(slack_message, "action: {action}")
-         
 
         title = payload.get("pull_request", {}).get("title")
         if title:
@@ -100,12 +105,13 @@ class GithubParser:
                 slack_message, f"repository link: {repository_link}"
             )
 
-
         # TODO
         # save reviewer and author info
         # handle if switching base branch notification?
 
-        self.send_slack_message_to_channel(slack_message, send_slack_message=send_slack_message)
+        self.send_slack_message_to_channel(
+            slack_message, send_slack_message=send_slack_message
+        )
         return slack_message
 
     def parse_workflow_run(self, payload, send_slack_message=True):
@@ -126,13 +132,13 @@ class GithubParser:
         slack_message = SlackClient.add_to_slack_string(
             slack_message, "webhook type: workflow run"
         )
-        status = payload.get('workflow_run', {}).get('status')
-        if status and status != '':
+        status = payload.get("workflow_run", {}).get("status")
+        if status and status != "":
             slack_message = SlackClient.add_to_slack_string(
                 slack_message,
                 f"status: {status}",
             )
-        conclusion = payload.get('workflow_run', {}).get('conclusion')
+        conclusion = payload.get("workflow_run", {}).get("conclusion")
         if conclusion and conclusion != "":
             slack_message = SlackClient.add_to_slack_string(
                 slack_message, f"conclusion: {conclusion}"
@@ -141,7 +147,9 @@ class GithubParser:
             slack_message, f"url: {payload.get('workflow', {}).get('html_url')}"
         )
 
-        self.send_slack_message_to_channel(slack_message, send_slack_message=send_slack_message)
+        self.send_slack_message_to_channel(
+            slack_message, send_slack_message=send_slack_message
+        )
         return slack_message
 
     def parse_workflow_job(self, payload, send_slack_message=True):
@@ -162,13 +170,13 @@ class GithubParser:
         slack_message = SlackClient.add_to_slack_string(
             slack_message, "webhook type: workflow job"
         )
-        status = payload.get('workflow_job', {}).get('status')
-        if status and status != '':
+        status = payload.get("workflow_job", {}).get("status")
+        if status and status != "":
             slack_message = SlackClient.add_to_slack_string(
                 slack_message,
                 f"status: {status}",
             )
-        conclusion = payload.get('workflow_job', {}).get('conclusion')
+        conclusion = payload.get("workflow_job", {}).get("conclusion")
         if conclusion and conclusion != "":
             slack_message = SlackClient.add_to_slack_string(
                 slack_message, f"conclusion: {conclusion}"
@@ -178,5 +186,7 @@ class GithubParser:
             f"url: {payload.get('workflow_job', {}).get('html_url')}",
         )
 
-        self.send_slack_message_to_channel(slack_message, send_slack_message=send_slack_message)
+        self.send_slack_message_to_channel(
+            slack_message, send_slack_message=send_slack_message
+        )
         return slack_message